Made function for finding robot pos and one for robot heading. Refactor grid-converter
