+++ src/on_computer/ComputerMain.py
@@ -1,4 +1,6 @@
 import socket
+
+from cv2 import filterHomographyDecompByVisibleRefpoints
 import PathfindingAlgorithm
 from PathfindingAlgorithm import grid, Node
 import json
@@ -6,46 +8,33 @@ from ArrayGenerator import create_border_array
 from computer_vision.ComputerVision import get_grid, get_masks_from_camera, get_robot_pos_with_mask
 from computer_vision.Displacement import move_point
 from computer_vision.Robot_direction import calculate_heading
+from Convert_to_node_grid import convert_to_grid
 
 masks = get_masks_from_camera()
 
 raw_grid_data = get_grid(masks['red'], masks['orange'], masks['white'])
 ball_node = PathfindingAlgorithm.node
-
-def convert_to_grid(data):
-    rows = len(data)
-    cols = len(data[0])
-    grid = []
-
-    for i in range(rows):
-        row_nodes = []
-        for j in range(cols):
-            node = Node(grid, j, i)
-            element = data[i][j]
-            if element == 1:
-                node.type = 'wall'
-            elif element == 0:
-                node.type = 'road'
-            elif element == 2:
-                node.type = 'ball'
-            row_nodes.append(node)
-        grid.append(row_nodes)
-
-    return grid
-
 grid = convert_to_grid(raw_grid_data)
 
 # below, y is first and x is second as the grid is a matrix not a cartesian plane
 
-robot_head = get_robot_pos_with_mask(masks['blue'])
-robot_real_position = move_point(robot_head, grid)
-robot_tail = get_robot_pos_with_mask(masks['green'])
-robot_real_tail = move_point(robot_tail,grid)
-robot_angle = calculate_heading(robot_real_tail,robot_real_position)
-print(robot_angle)
 
+def find_start_node():
+    robot_camera_position = get_robot_pos_with_mask(masks['blue'])
+    robot_real_position = move_point(robot_camera_position, grid) #Accounting for displacement
+    if (robot_real_position is None):
+        print("Robot position is not found")
+        return
+    return grid[int(robot_real_position[0])][int(robot_real_position[1])]
 
-start_node = grid[int(robot_real_position[0])][int(robot_real_position[1])]
+def get_robot_angle():
+    robot_camera_tail = get_robot_pos_with_mask(masks['green'])
+    robot_real_tail = move_point(robot_camera_tail,grid)
+    robot_angle = calculate_heading(robot_real_tail , find_start_node() )
+    if robot_angle is None:
+        print("Robot angle is not found")
+        return
+    return robot_angle
 
 def find_first_ball(grid):
     for row in grid:
@@ -55,6 +44,7 @@ def find_first_ball(grid):
     return None
 
 
+start_node = find_start_node()
 end_node = find_first_ball(grid)
 
 print(start_node.x, start_node.y)
