+++ src/on_computer/computer_vision/Displacement.py
@@ -41,9 +41,9 @@ def find_displacement_of_robot(grid, robot_field_position):
     return displacement
 
 def move_point(robot_position, grid):
-    if robot_position == None or grid == None:
+    if  robot_position is None or grid is None:
         print("Robot position or grid is not found")
-        return None
+        return
 
     # Convert points to numpy arrays
     distance = find_displacement_of_robot(grid, robot_position)
